requirements.txt
faiss-cpu
sentence-transformers
numpy
pandas
streamlit
